thread ids are using int

+++ src/main/java/com/dynatrace/openkit/providers/ThreadIDProvider.java
@@ -25,5 +25,5 @@ public interface ThreadIDProvider {
      *
      * @return
      */
-    long getThreadID();
+    int getThreadID();
 }
